add backend

+++ backend-typescript/src/authorization/resource/resource.orchestrator.ts
@@ -8,15 +8,20 @@ import {getSortOrderOrDefault} from "../../sort.order.util";
 import {ResourcePermission} from "../resource-permission/resource.permission";
 import {ResourcePermissionRepository} from "../resource-permission/resource.permission.repository";
 import {createResourcePermissionRepositoryFactory} from "../resource-permission/resource.permission.repository.factory";
+import {ResourceTypeRepository} from "../resource-type/resource.type.repository";
+import {createResourceTypeRepositoryFactory} from "../resource-type/resource.type.repository.factory";
+import {ResourceType} from "../resource-type/resource.type";
 
 export class ResourceOrchestrator {
 
   private resourceRepository:ResourceRepository;
   private resourcePermissionRepository: ResourcePermissionRepository;
+  private resourceTypeRepository:ResourceTypeRepository;
 
   constructor() {
     this.resourceRepository = createResourceRepositoryFactory();
     this.resourcePermissionRepository = createResourcePermissionRepositoryFactory();
+    this.resourceTypeRepository = createResourceTypeRepositoryFactory();
   }
 
   getResourcesByArray(number:number, size:number, field:string, direction:string, assignedArray:string[]):Observable<Result<any>> {
@@ -47,22 +52,37 @@ export class ResourceOrchestrator {
       });
   }
   getResources(number:number, size:number, field:string, direction:string):Observable<Result<any>> {
-    let sort:string = getSortOrderOrDefault(field, direction);
-    return this.resourceRepository
-      .getResources(number, size, sort)
-      .flatMap(value => {
-        return this.resourceRepository
-          .getResourceCount()
-          .map(count => {
-            let shapeResourcesResp:any = shapeResourcesResponse(value, number, size, value.length, count, sort);
-            return new Result<any>(false, "resources", shapeResourcesResp);
+    let sort: string = getSortOrderOrDefault(field, direction);
+    return this.resourceRepository.getResources(number, size, sort)
+      .switchMap((resources: Resource[]) => {
+        if (resources.length === 0) {
+          let shapeResourcesResp: any = shapeResourcesResponse(resources, 0, 0, 0, 0, sort);
+          return Observable.of(new Result<any>(false, "no data found", shapeResourcesResp));
+        }else {
+          let resourceTypeIds:string[] = resources.map(x => { if(x.resourceTypeId) return x.resourceTypeId});
+          return this.resourceTypeRepository.getResourceTypeByIds(resourceTypeIds)
+            .switchMap((resourceTypes: ResourceType[]) => {
+              resources.forEach(value => {
+                let index = resourceTypes.findIndex(x => x.resourceTypeId === value.resourceTypeId);
+                value.resourceType = resourceTypes[index];
+              });
+              return this.resourceRepository
+                .getResourceCount()
+                .map(count => {
+                  let shapeResourcesResp:any = shapeResourcesResponse(resources, number, size, resources.length, count, sort);
+                  return new Result<any>(false, "resources", shapeResourcesResp);
+                });
           });
+        }
       });
   };
 
   addResource(resource:Resource, resourcePermissions: ResourcePermission[]):Observable<Resource> {
     return this.resourceRepository.addResource(resource)
       .switchMap( value => {
+        if(resourcePermissions.length === 0) {
+          return Observable.of(value);
+        }
         let resourceId = value.resourceId;
         if(resourceId) {
           resourcePermissions.forEach(val => {
@@ -79,12 +99,24 @@ export class ResourceOrchestrator {
   };
 
   getResourceById(resourceId:string):Observable<Resource> {
-    return this.resourceRepository.getResourceById(resourceId);
+    return this.resourceRepository.getResourceById(resourceId)
+      .switchMap((resource:Resource) => {
+        return this.resourceTypeRepository.getResourceTypeById(resource.resourceTypeId)
+          .map(resourceType => {
+          if(resourceType) {
+            resource.resourceType = resourceType
+          }
+          return resource;
+        });
+      });
   };
 
   updateResource(resourceId:string, resource:Resource, resourcePermissions: ResourcePermission[]):Observable<number> {
     return this.resourceRepository.updateResource(resourceId, resource)
       .switchMap( numReplaced => {
+        if(resourcePermissions.length === 0) {
+          return Observable.of(numReplaced);
+        }
         if(numReplaced) {
           return this.resourcePermissionRepository.deleteResourcePermission(resourceId)
             .switchMap(numReplaced2 => {
